requirements.txt
sqlalchemy>=2.0.0
faker>=15.0.0
pydantic>=2.0.0
loguru>=0.7.3
requests>=2.0.0
python-dotenv>=0.19.0
ipykernel>=7.1.0
ipython>=8.38.0
jupyterlab>=4.5.3
notebook>=7.5.3
pandas>=2.3.3
scipy>=1.15.3
matplotlib>=3.10.8
seaborn>=0.13.2
pymysql>=1.1.0
streamlit>=1.28.0
mysql-connector-python>=8.0.33